refactor(codemods): deduplicate decorator collection and field extraction helpers
